% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex31.R
\docType{data}
\name{ex31}
\alias{ex31}
\title{ex31 from Duchateau, L. and Janssen, P. and Rowlands, G. J. (1998).\emph{Linear Mixed Models. An Introduction with applications in Veterinary Research}. International Livestock Research Institute.}
\format{A \code{data.frame} with 38 rows and 6 variables.}
\usage{
data(ex31)
}
\description{
ex31 is.
}
\details{
\itemize{
       \item herd      Herds of Cattle
       \item animal_id Animal ID
       \item PCV1      Packed Cell Volume (PCV) determined at the time of treatment
       \item PCV2      Packed Cell Volume (PCV) determined at a month later following treatment
       \item dose      Dose of Drugs
       \item drug      Two drugs against trypanosomosis, Berenil and Samorin, are studied
       }
}
\examples{
data(ex31)
}
\references{
\enumerate{
\item Duchateau, L. and Janssen, P. and Rowlands, G. J. (1998).\emph{Linear Mixed Models. An Introduction with applications in Veterinary Research}.
             International Livestock Research Institute.
 }
}
\seealso{
\code{\link{Examp1.3.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         }
}
\keyword{datasets}
