% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex33.R
\docType{data}
\name{ex33}
\alias{ex33}
\title{ex33 from Duchateau, L. and Janssen, P. and Rowlands, G. J. (1998).\emph{Linear Mixed Models. An Introduction with applications in Veterinary Research}. International Livestock Research Institute.}
\format{A \code{data.frame} with 40 rows and 4 variables.}
\usage{
data(ex33)
}
\description{
ex33 is.
}
\details{
\itemize{
       \item herd two treatment 0 and 1
       \item drug unit of observation or observation ID
       \item dose  is continuous \& may be assumed Gaussian
       \item PCVDif  is the number of "successes"(N and F specify a binomial response)
       }
}
\examples{
data(ex33)
}
\references{
\enumerate{
\item Duchateau, L. and Janssen, P. and Rowlands, G. J. (1998).\emph{Linear Mixed Models. An Introduction with applications in Veterinary Research}.
             International Livestock Research Institute.
 }
}
\seealso{
\code{\link{Examp1.3.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         }
}
\keyword{datasets}
